catch-all/06_bots_telegram/05_rss_bot/requirements.txt
python-dotenv==1.0.1
feedparser==6.0.11
python-telegram-bot[job-queue]==21.4
